refactor(source): abstract components and improve source code
---
Massive rework including creation of barrel files for Portfolio assets, and improvement of the source code of the `Home` and `Contact` components, as well as their sub-components, for better readability and general organization. Finally, abstracted the components:

* `Introduction`
* `ContactForm`
* `Modal`
* `RotatingLogo`
* `CardsMarquee`

and the utilities:

* `customFormats` › `PhoneNumber()`
* `customTypes` › `DataStructure`
