+++ src/front/assets/img/Portfolio/Busquedas/index.js
@@ -1,0 +1,2 @@
+export { default as cardBg } from "./cardBg.jpg";
+export { default as logo } from "./logo.png";
