+++ src/front/assets/img/Portfolio/index.js
@@ -1,0 +1,6 @@
+import * as busquedas from "./Busquedas";
+import * as cosmonauta from "./Cosmonauta";
+import * as jp from "./JPPortfolio";
+import * as mura from "./Mura";
+
+export default { busquedas, cosmonauta, jp, mura };
