+++ src/front/pages/Contact.jsx
@@ -1,6 +1,6 @@
 import { CardsContact } from "../components/Contact/CardsContact"
 import { Afterwards } from "../components/Contact/Afterwards"
-import HeaderContact from "../components/HeaderContact"
+import { HeaderContact } from "../components/HeaderContact"
 import { SEO } from "../utils/SEO"
 
 export const Contact = () => {
