+++ src/front/utils/customTypes.js
@@ -1,0 +1,19 @@
+/** ### Form Data Structure
+ * 
+ * @typedef DataStructure
+ * @prop {string[]} needs
+ * @prop {number} stage
+ * @prop {string} problemDescription
+ * @prop {number} idealTimeframe
+ * @prop {PersonalData} personalData
+ * @prop {boolean} newsletterOptIn
+ * 
+ * @typedef PersonalData
+ * @prop {string} wholeName
+ * @prop {string} email
+ * @prop {string} phone
+ * @prop {string} projectName
+ * 
+*/
+
+export default {};
